refactor(torghut): remove dead service scaffolding
