+++ services/torghut/app/trading/decisions/decision_engine_runtime_methods.py
@@ -325,7 +325,6 @@ class DecisionEngine(
     _DecisionEngineFields,
     _DecisionEngineCoreMethods,
     _DecisionEngineRuntimeMethods,
-    object,
 ):
     pass
