+++ services/torghut/app/trading/execution/order_executor_submission_methods.py
@@ -478,7 +478,6 @@ class OrderExecutor(
     _OrderExecutorFields,
     _OrderExecutorCoreMethods,
     _OrderExecutorSubmissionMethods,
-    object,
 ):
     pass
