+++ services/torghut/app/trading/ingest/clickhouse_signal_ingestor_persistence_methods.py
@@ -147,7 +147,6 @@ class ClickHouseSignalIngestor(
     _ClickHouseSignalIngestorCoreMethods,
     _ClickHouseSignalIngestorMarketMethods,
     _ClickHouseSignalIngestorPersistenceMethods,
-    object,
 ):
     pass
