+++ services/torghut/app/trading/scheduler/governance/governance_mixin_runtime_methods.py
@@ -496,7 +496,6 @@ class TradingSchedulerGovernanceMixin(
     _TradingSchedulerGovernanceLifecycleMethods,
     _TradingSchedulerGovernanceDecisionMethods,
     _TradingSchedulerGovernanceRuntimeMethods,
-    object,
 ):
     pass
