+++ services/torghut/app/trading/scheduler/paper_route_probe/mixin.py
@@ -9,7 +9,6 @@ from .retry_decisions import SimplePipelinePaperRouteProbeRetryDecisionMixin
 class SimplePipelinePaperRouteProbeMixin(
     SimplePipelinePaperRouteProbeRetryDecisionMixin,
     SimplePipelinePaperRouteProbeProcessingMixin,
-    object,
 ):
     pass
