+++ services/torghut/app/trading/scheduler/pipeline/trading_pipeline.py
@@ -20,7 +20,6 @@ class TradingPipeline(
     TradingPipelineRuntimeGatesMixin,
     TradingPipelineReviewMixin,
     TradingPipelineReviewOutcomeMixin,
-    object,
 ):
     pass
