+++ services/torghut/app/trading/scheduler/submission_preparation/direct_submission.py
@@ -792,6 +792,5 @@ class SimplePipelineSubmissionPreparationMixin(
     SimplePipelineSubmissionQuoteSizingMixin,
     SimplePipelineSubmissionQuoteRouteabilityMixin,
     SimplePipelineDirectSubmissionMixin,
-    object,
 ):
     pass
