+++ services/torghut/app/whitepapers/workflow/whitepaper_workflow_service.py
@@ -48,7 +48,6 @@ class WhitepaperWorkflowService(
     _WhitepaperWorkflowVerdictMethods,
     _WhitepaperWorkflowRolloutMethods,
     _WhitepaperWorkflowApiMethods,
-    object,
 ):
     pass
